apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.5
pydantic==2.7.1
pydantic-settings==2.2.1
boto3==1.34.100
botocore==1.34.100
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
pyyaml==6.0.1
celery==5.4.0
redis==5.0.4
python-multipart==0.0.9
pdfplumber==0.11.0
anthropic==0.40.0
voyageai==0.3.2
